Move parser classes to diagostics.ts
