+++ src/cmake.ts
@@ -15,7 +15,14 @@ import {FileDiagnostic,
         DiagnosticParser,
         LineParseResult,
         LineParseStatus,
-        PARSER_NEEDS_MORE} from './diagnostics';
+        PARSER_NEEDS_MORE,
+        PARSER_FAIL,
+        CMakeDiagnosticParser,
+        GCCDiagnosticParser,
+        GNULDDiagnosticParser,
+        MSVCDiagnosticParser,
+        GHSDiagnosticParser,
+        } from './diagnostics';
 import {util} from './util';
 
 type Maybe<T> = util.Maybe<T>;
@@ -447,276 +454,6 @@ class CMakeTargetListParser extends util.OutputParser {
                 : l)
             .map(l => / /.test(l) ? l.substr(0, l.indexOf(' ')) : l)
             .map(l => l.replace(':', ''));
-    }
-}
-
-class CMakeDiagnosticParser extends DiagnosticParser {
-    private _cmakeDiag: Maybe<FileDiagnostic>;
-
-    constructor(binaryDir: string) {
-        super(binaryDir);
-        this._cmakeDiag = null;
-    }
-
-    public parseLine(line: string): LineParseResult {
-        if (!line) {
-            const diagnostic = this._cmakeDiag;
-            this._cmakeDiag = null;
-            if (diagnostic) {
-                return {
-                    status: LineParseStatus.Done,
-                    diagnostic
-                };
-            } else {
-                return PARSER_NEEDS_MORE;
-            }
-        }
-
-        const cmake_re = /CMake (.*?) at (.*?):(\d+) \((.*?)\):\s*(.*)/;
-        const res = cmake_re.exec(line);
-        if (!res) {
-            if (this._cmakeDiag) {
-                this._cmakeDiag.diag.message = '\n' + line;
-                return { status: LineParseStatus.Done, diagnostic: null };
-            }
-            return PARSER_NEEDS_MORE;
-        }
-
-        const [full, level, filename, linestr, command, what] = res;
-        if (!filename || !linestr || !level) {
-            if (this._cmakeDiag) {
-                this._cmakeDiag.diag.message = '\n' + line;
-                return { status: LineParseStatus.Done, diagnostic: null };
-            }
-            return { status: LineParseStatus.NeedMore };
-        }
-
-        this._cmakeDiag = <FileDiagnostic>{};
-        this._cmakeDiag.filepath = path.isAbsolute(filename)
-            ? filename
-            : path.join(vscode.workspace.rootPath, filename);
-        this._cmakeDiag.key = full;
-        const lineNr = Number.parseInt(linestr) - 1;
-
-        this._cmakeDiag.diag = new vscode.Diagnostic(
-            new vscode.Range(
-                lineNr,
-                0,
-                lineNr,
-                Number.POSITIVE_INFINITY
-            ),
-            what,
-            {
-                "Warning": vscode.DiagnosticSeverity.Warning,
-                "Error": vscode.DiagnosticSeverity.Error,
-            }[level]
-        );
-        this._cmakeDiag.diag.source = 'CMake (' + command + ')';
-        return { status: LineParseStatus.Done, diagnostic: null };
-    }
-}
-
-/**
- * Parses a diagnostic message from GCC
- *
- * @class GCCDiagnosticParser
- * @extends {DiagnosticParser}
- */
-class GCCDiagnosticParser extends DiagnosticParser {
-    public parseLine(line: string): LineParseResult {
-        const diag = diagnostics.parseGCCDiagnostic(line);
-        if (!diag) {
-            return { status: LineParseStatus.NeedMore };
-        }
-        const abspath = path.isAbsolute(diag.file)
-            ? diag.file
-            : path.normalize(path.join(this.binaryDir, diag.file));
-        const vsdiag = new vscode.Diagnostic(
-            new vscode.Range(diag.line, diag.column, diag.line, diag.column),
-            diag.message,
-            {
-                error: vscode.DiagnosticSeverity.Error,
-                warning: vscode.DiagnosticSeverity.Warning,
-                note: vscode.DiagnosticSeverity.Information,
-            }[diag.severity]
-        );
-        vsdiag.source = 'GCC';
-        return {
-            status: LineParseStatus.Done,
-            diagnostic: {
-                filepath: abspath,
-                key: diag.full,
-                diag: vsdiag,
-            }
-        };
-    }
-}
-
-/**
- * Parses a diagnostic message from GCC
- *
- * @class GNULDDiagnosticParser
- * @extends {DiagnosticParser}
- */
-class GNULDDiagnosticParser extends DiagnosticParser {
-    public parseLine(line: string): LineParseResult {
-        const diag = diagnostics.parseGNULDDiagnostic(line);
-        if (!diag) {
-            return { status: LineParseStatus.NeedMore };
-        }
-        const abspath = path.isAbsolute(diag.file) ? diag.file : path.normalize(path.join(this.binaryDir, diag.file));
-        const vsdiag = new vscode.Diagnostic(
-            new vscode.Range(diag.line, 0, diag.line, Number.POSITIVE_INFINITY),
-            diag.message,
-            {
-                error: vscode.DiagnosticSeverity.Error,
-                warning: vscode.DiagnosticSeverity.Warning,
-                note: vscode.DiagnosticSeverity.Information,
-            }[diag.severity]
-        );
-        vsdiag.source = 'Link';
-        return {
-            status: LineParseStatus.Done,
-            diagnostic: {
-                filepath: abspath,
-                key: diag.full,
-                diag: vsdiag,
-            }
-        };
-    }
-}
-
-/**
- * Parses an MSVC diagnostic.
- *
- * @class MSVCDiagnosticParser
- * @extends {DiagnosticParser}
- */
-class MSVCDiagnosticParser extends DiagnosticParser {
-    /**
-     * @brief Obtains a reference to a TextDocument given the name of the file
-     */
-    private getTextDocumentByFileName(file: string): Maybe<vscode.TextDocument> {
-        const documents = vscode.workspace.textDocuments;
-        let document: Maybe<vscode.TextDocument> = null;
-        if (documents.length !== 0) {
-            const filtered = documents.filter((doc: vscode.TextDocument) => {
-                return doc.fileName.toUpperCase() === file.toUpperCase();
-            });
-            if (filtered.length !== 0) {
-                document = filtered[0];
-            }
-        }
-        return document;
-    }
-
-    /**
-     * @brief Gets the range of the text of a specific line in the given file.
-     */
-    private getTrimmedLineRange(file: string, line: number): vscode.Range {
-        const document = this.getTextDocumentByFileName(file);
-        if (document && (line < document.lineCount)) {
-            const text = document.lineAt(line).text + '\n';
-            let start = 0;
-            let end = text.length - 1;
-            let is_space = (i) => { return /\s/.test(text[i]); };
-            while ((start < text.length) && is_space(start))++start;
-            while ((end >= start) && is_space(end))--end;
-
-            return new vscode.Range(line, start, line, end);
-        } else
-            return new vscode.Range(line, 0, line, 0);
-    }
-
-    public parseLine(line: string): LineParseResult {
-        const msvc_re = /^\s*(?!\d+>)\s*([^\s>].*)\((\d+|\d+,\d+|\d+,\d+,\d+,\d+)\):\s+(error|warning|info)\s+(\w{1,2}\d+)\s*:\s*(.*)$/;
-        const res = msvc_re.exec(line);
-        if (!res)
-            return PARSER_NEEDS_MORE;
-        const full = res[0];
-        const file = res[1];
-        const location = res[2];
-        const severity = res[3];
-        const code = res[4];
-        const message = res[5];
-        const abspath = path.isAbsolute(file)
-            ? file
-            : path.normalize(path.join(this.binaryDir, file));
-        const loc = (() => {
-            const parts = location.split(',');
-            if (parts.length === 1)
-                return this.getTrimmedLineRange(file, Number.parseInt(parts[0]) - 1);
-            if (parts.length === 2)
-                return new vscode.Range(
-                    Number.parseInt(parts[0]) - 1,
-                    Number.parseInt(parts[1]) - 1,
-                    Number.parseInt(parts[0]) - 1,
-                    Number.parseInt(parts[1]) - 1
-                );
-            if (parts.length === 4)
-                return new vscode.Range(
-                    Number.parseInt(parts[0]) - 1,
-                    Number.parseInt(parts[1]) - 1,
-                    Number.parseInt(parts[2]) - 1,
-                    Number.parseInt(parts[3]) - 1
-                );
-            throw new Error('Unable to determine location of MSVC error');
-        })();
-        const diag = new vscode.Diagnostic(
-            loc,
-            message,
-            {
-                error: vscode.DiagnosticSeverity.Error,
-                warning: vscode.DiagnosticSeverity.Warning,
-                info: vscode.DiagnosticSeverity.Information,
-            }[severity]
-        );
-        diag.code = code;
-        diag.source = 'MSVC';
-        return {
-            status: LineParseStatus.Done,
-            diagnostic: {
-                filepath: abspath,
-                key: full,
-                diag: diag,
-            }
-        };
-    }
-}
-
-/**
- * Parses a diagnostic message from Green Hills Compiler.
- * Use single line error reporting when invoking GHS compiler (--no_wrap_diagnostics --brief_diagnostics).
- *
- * @class GHSDiagnosticParser
- * @extends {DiagnosticParser}
- */
-class GHSDiagnosticParser extends DiagnosticParser {
-    public parseLine(line: string): LineParseResult {
-        const diag = diagnostics.parseGHSDiagnostic(line);
-        if (!diag)
-            return PARSER_NEEDS_MORE;
-        const abspath = path.isAbsolute(diag.file)
-            ? diag.file
-            : path.normalize(path.join(this.binaryDir, diag.file));
-        const vsdiag = new vscode.Diagnostic(
-            new vscode.Range(diag.line, diag.column, diag.line, diag.column),
-            diag.message,
-            {
-                error: vscode.DiagnosticSeverity.Error,
-                warning: vscode.DiagnosticSeverity.Warning,
-                remark: vscode.DiagnosticSeverity.Information,
-            }[diag.severity]
-        );
-        vsdiag.source = 'GHS';
-        return {
-            status: LineParseStatus.Done,
-            diagnostic: {
-                filepath: abspath,
-                key: diag.full,
-                diag: vsdiag,
-            }
-        };
     }
 }
